linode network - layout changes and other tweaks
