+++ test/data.js
@@ -11,7 +11,9 @@ export const testLinode = {
     },
     private: {
       link_local: 'fe80::f03c:91ff:fe96:43e7',
-      ipv4: [],
+      ipv4: [
+        '321.456.789.1',
+      ],
     },
   },
   id: 'linode_1234',
